+++ app/assets/CHANGELOG.md
@@ -1,6 +1,10 @@
 ## 1.15.1 (unreleased)
 
+- feat: support Internet Explorer 8 (IE8) in web share (@Tienisto)
+- feat: save auto accept state when switching encryption mode in web share (@Tienisto)
+- fix: count PIN tries correctly in web share (@Tienisto)
 - fix(windows): crash when sum of file sizes greater than 2 GB (@Tienisto)
+- fix(windows): bundle required DLLs files to avoid crash on app start (@Tienisto)
 
 ## 1.15.0 (2024-07-15)